requirements.txt
Pillow==9.2.0
PyAutoGUI==0.9.53
numpy==1.23.4
opencv_python==4.6.0.66
pytesseract==0.3.10
requests==2.28.1
pywin32==304
psutil==5.9.2
python-socketio[client]==5.7.2
